Refactor the whole structure of the library
Now things specific to the bundled DefaultLayout is moved to
`FrameKitCLILib` and `DefaultFrameKitLayout`. For example,
`ContentProvider` was removed to stop constraint to `Content`.
`Content` now became `DefaultContent` in `DefaultFrameKitLayout` so that
you can include any types of content as you wish.

There is also a rename to make the terminalogy consistent.
`AppScreenshotView` is now called `StoreScreenshotView`.